BlockClass.py
Code from when I first created this repository to when I created BlockClass1.
