--- Block Class.py
@@ -1,12 +1,0 @@
-class Block:
-    def __init__(self, x, y, width, height, IDGround, IDAir):
-        self.x = x
-        self.y = y
-        self.width = width
-        self.height = height
-        self.IDGround = IDGround
-        self.IDAir = IDAir
-
-block = Block(1,1,100,100,0,1)
-
-#Test Commit
